--- exp1/backwardMatchSeg.py
@@ -1,102 +1,0 @@
-text = "　北京时间2月20日，2017年NBA全明星正赛开始之前的训练中出现了有趣的\
-一幕，起初，威少脱离其他西部全明星球员而独自练习投篮。但是在全队跑篮的过程中，\
-威少加入了西部全明星队，并且还和凯文-杜兰特出现了一些互动。从推特上媒体拍的照片\
-来看，西部全明星球员的赛前投篮热身中，威少一个人在一个半场投篮，而其他西部\
-全明星球员则在另一个半场投篮。看来对于杜兰特，威少开始是拒绝的。\
-　在外界看来，威少是在刻意避开杜兰特。也许是看威少一个人投篮太孤独，\
-詹姆斯-哈登随后加入了威少。\
-　然而在之后的全队跑篮中，威少加入了西部全明星队，而且还和杜兰特出现了\
-大家期待的互动，他们互相传球了！\
-　据了解，杜兰特还为威少送出上篮助攻。而勇士记者安东尼-斯莱特在推特上表示：\
-“虽然威少和杜兰特并没有变得友好，但是他们也没有忽视对方：KD和威少互相为对方\
-传球和抢篮板。”\
-　“天呐，天呐，看见了没？杜兰特的更衣柜刚好在韦斯布鲁克更衣柜的对面。”梅拉斯\
-这样写道。随后他还发布了一段小视频，可以清楚的看出，韦斯布鲁克与杜兰特之间，\
-相隔了7名球员。\
-　在韦斯布鲁克身旁的，是身披15号的考辛斯，而杜兰特则位于对角位置。他的左侧位置\
-是汤普森、库里以及格林。勇士四人组的更衣柜位置，是挨在一块儿的。\
-　尽管昔日同在雷霆效力，但杜韦二少如今关系非常恶劣。勇士客场击败雷霆赛后，\
-两人先后进入一家牛排馆吃饭，相隔五米，全程毫无交流。到了全明星周末期间，\
-韦斯布鲁克与杜兰特也曾迎面碰到过。不过两人都装作没有看见，宛若对方是空气一般。\
-　破裂的感情如何去修复？或许只能依靠时间。有趣的是，在昨天的三分球大赛上，现场\
-的主持人在介绍到场嘉宾时出现口误，将端坐在场边的杜兰特称为“来自俄克拉荷马”，\
-为此杜兰特显得一脸不爽。\
-　本届全明星正赛，最大的看点就是杜韦二少的重逢会擦出怎样的火花，而这一切都掌握\
-在勇士主帅史蒂夫-科尔的手里。同时，韦少能否完成全明星MVP的三连庄成为了这次正赛\
-的另一个噱头。张卫平指导对本次正赛也有他自己的看法，让我们看看他是怎么说的。\
-　张卫平预测全明星正赛：3因素助韦少MVP三连庄\
-　张卫平指导预测全明星正赛结果\
-　对于全明星赛MVP的归属，张指导认为韦少当选的机会很大。韦少已经连续两年成为\
-全明星MVP的归属。（2015年全明星赛韦少轰下41分夺得全明星赛MVP。2016年全明星赛\
-他又砍下31分蝉联了全明星赛MV-P。韦少成为了继佩蒂特之后NBA历史上第二位蝉联\
-全明星赛MVP的球员。）如果韦少今年再夺全明星赛MVP，那么他将成为NB-A史上第一\
-位完成全明星赛MVP三连庄的球员，因此他将为这个目标全力冲击。\
-　为什么这么说呢？首先，前凯尔特人队，现快船队主教练道格-里弗斯向科尔提建议，\
-要让自己的死敌在全明星赛上疯狂地发挥，免得将来和他见面的时候他和你玩命。现在\
-勇士在西部的死敌就是韦少带领的雷霆。究其原因，还是因为杜兰特和韦少之间冗长的\
-肥皂剧。因此，科尔一定会让韦少在全明星赛上打个痛快。\
-　其次，张指导认为，勇士四巨头不会因为科尔在全明星赛上控制他们的出场时间而\
-埋怨自己的主教练，但你的对手会因为你对他的怠慢而记下仇来。因此，科尔不出意外\
-应该会听从里弗斯的建议，让韦少玩命打。\
-　第三点，从韦少自己的性格来看，虽然全明星赛是个娱乐性质的比赛，没有人特别\
-认真地去打，但是韦少不会在意这些的，他的特点是上场就干，这就是他的性格。两次\
-全明星MVP就是最好的佐证，大伙也心照不宣。而且韦少目前虽然保持着场均三双，但却\
-不是全明星先发，这让他在心里一直憋着一口气证明自己。\
-　如此看来，如果科尔选择让韦少战个痛快，再加上韦少自己的性格，他的第三座全明星\
-MVP应该是稳稳地收下了。但是张指导说了，不能说韦少一定能拿MVP，这只是个预测，\
-具体的结果还是要看临场的发挥。（栗旬）"
-
-
-def backward_max_match(sentence, dictionary, max_word_length):
-    index = len(sentence)
-    result = []
-    while index > 0:
-        word = None
-        for size in range(max_word_length, 0, -1):
-            if index - size < 0: continue
-            piece = sentence[index - size:index]
-            if piece in dictionary:
-                word = piece
-                result.append(word)
-                index -= size
-                break
-        if word is None:
-            word = sentence[index - 1]
-            result.append(word)
-            index -= 1
-    return result[::-1]  # 反转结果列表以保持原句的顺序
-
-
-# 示例词典，实际使用时请扩充以包含更多词汇
-dictionary = {'北京', 'NBA', '全明星', '训练', '有趣', '一幕', '起初', '威少', '脱离', '其他', '西部', '球员', '独自',
-              '练习', '投篮', '全队', '跑篮', '过程中', '加入', '队', '并且', '凯文-杜兰特', '出现', '一些', '互动',
-              '推特', '媒体', '拍', '照片', '赛前', '热身', '一个人', '一个', '半场', '看来', '对于', '开始', '拒绝',
-              '外界', '刻意', '避开', '孤独', '詹姆斯-哈登', '随后', '之后', '大家', '期待', '他们', '互相', '传球',
-              '了解', '送出', '上篮', '助攻', '勇士', '记者', '安东尼-斯莱特', '表示', '没有', '变得', '友好', '但是',
-              '也', '没有', '忽视', '对方', 'KD', '抢篮板', '天呐', '看见', '没', '更衣柜', '刚好', '韦斯布鲁克',
-              '对面', '梅拉斯', '写道', '发布', '小视频', '清楚', '相隔', '名球员', '身旁', '考辛斯', '位于', '对角',
-              '位置', '左侧', '汤普森', '库里', '格林', '四人组', '挨在', '一块', '尽管', '昔日', '同在', '雷霆',
-              '效力', '关系', '非常', '恶劣', '客场', '击败', '赛后', '进入', '牛排馆', '吃饭', '相隔', '五米', '全程',
-              '毫无', '交流', '全明星周末', '期间', '迎面', '碰到', '装作', '看见', '宛若', '空气', '破裂', '感情',
-              '修复', '依靠', '时间', '三分球', '大赛', '现场', '主持人', '介绍', '到场', '嘉宾', '口误', '端坐',
-              '场边', '俄克拉荷马', '不爽', '本届', '看点', '重逢', '火花', '掌握', '主帅', '史蒂夫-科尔', '手里',
-              '同时', '能否', '完成', 'MVP', '三连庄', '成为', '这次', '正赛', '另一个', '噱头', '张卫平', '指导',
-              '预测', '结果', '归属', '认为', '当选', '机会', '很大', '已经', '连续', '两年', '成为', '（', '）', '轰下',
-              '分夺得', '砍下', '蝉联', '成为了', '继', '之后', '第二位', '如果', '今年', '再', '夺得', '那么', '将',
-              '第一位', '因此', '全力', '冲击', '为什么', '这么', '说', '首先', '前', '凯尔特人队', '快船队', '主教练',
-              '道格-里弗斯', '向', '提建议', '死敌', '疯狂', '地', '发挥', '免得', '将来', '见面', '时候', '玩命',
-              '现在', '西部', '雷霆', '究其', '原因', '还是', '因为', '冗长', '肥皂剧', '一定', '让', '打', '个',
-              '痛快', '其次', '四巨头', '不会', '因为', '控制', '出场', '时间', '埋怨', '教练', '对手', '怠慢', '记下',
-              '仇来', '不出', '意外', '应该', '听从', '建议', '玩命', '第三点', '性格', '娱乐性质', '认真', '上场',
-              '就是', '干', '这是', '佐证', '大家', '心照不宣', '目前', '保持', '场均', '三双', '不是', '先发', '心里',
-              '一直', '憋着', '一口气', '证明', '自己', '如此', '战个', '再加上', '应该', '稳稳', '地', '收下', '但是',
-              '不能', '说', '一定', '只是', '预测', '具体', '看', '临场', '发挥', '（栗旬）'}
-max_word_length = max(len(word) for word in dictionary)
-
-# 移除文本中的空格和特殊字符
-text = text.replace('\u3000', '').replace('\n', '')
-
-# 分词
-result = backward_max_match(text, dictionary, max_word_length)
-
-print(' / '.join(result))
